feat : client side role checking in create-course page test
