+++ lib/supabase.js
@@ -7,6 +7,14 @@ if (!SUPABASE_URL || !SUPABASE_ANON_KEY) {
   throw new Error("Supabase URL or Anon Key is missing!");
 }
 
-const supabase = createClient(SUPABASE_URL, SUPABASE_ANON_KEY);
+const supabase = createClient(SUPABASE_URL, SUPABASE_ANON_KEY, {
+  auth: {
+    cookieOptions: {
+      expiresIn: 60 * 60, 
+      sameSite: "Lax", 
+      secure: process.env.NODE_ENV === "production", 
+    },
+  },
+});
 
 export default supabase;
