+++ package.json
@@ -20,7 +20,6 @@
     "@supabase/ssr": "^0.5.2",
     "autoprefixer": "^10.4.20",
     "bcrypt": "^5.1.1",
-    "bootstrap": "^5.3.3",
     "cors": "^2.8.5",
     "node-releases": "^2.0.18",
   },
